Autotools work.  Samples still won't build.
There's a dependency issue.  When you read the output of ./configure,
you see that it's not able to find functions in the tdrmain library or
the tdrrenderer library.  Adding link directories and whatnot doesn't
seem to work.  I'm not an autotools expert, so I have no idea how to
remedy this situation.  Maybe going to CMake would be a better idea?