Moved the cube static code to the side.  Created new sample for models.
The Cube code is now a ::Cube() function in TDRMesh.  I don't want to remove
that code, as it's very simple and demonstrates what it takes to actually
have a model.  I may change that to a ::StaticModel(TDRMesh::Object) type
of function so that I can load other static objects to quickly get something
on the screen.