Added more functions for TDRMesh class.
This is the first step in making the TDRMesh class more sophisticated
than just loading a simple JSON file full of floats.  It will be able
to load a JSON file that contains several named objects.  There will be
a sample added that will test the TDRMesh class with various sample
object files.