Added tdrmeshtool
This will take collada objects and convert them to tdrmesh objects.
Obviously this will be a big undertaking.  The biggest task is to either
find a collada parser with few dependencies or to roll my own.  As an
example, assimp is 30MB gzipped...so it's way too big.  The quest continues.