Working on the XML parser.
Moving from JSON to XML.  I feel that if I make the actual buffer data
be a binary file that's simply pointed to by a "master" XML file, it'll
be much easier to manage.  The JSON parser was just too difficult to work
with large data sets.  But I also don't want to pull in another large
library into the source tree as a dependency or to put in source control
and maintain.